refactor: 패키지 구조 수정 & 불필요 파일 삭제
