--- app/src/main/java/woowacourse/shopping/domain/repository/CurrentProductRepository.kt
@@ -1,13 +1,0 @@
-package woowacourse.shopping.domain.repository
-
-import woowacourse.shopping.domain.model.RecentlyProduct
-
-interface CurrentProductRepository {
-    fun saveRecentlyProduct(recentlyProduct: RecentlyProduct)
-
-    fun getMostRecentlyProduct(): RecentlyProduct
-
-    fun getPagingRecentlyProduct(): List<RecentlyProduct>
-
-    fun removeRecentlyProduct(id: Long)
-}
