+++ app/src/main/java/woowacourse/shopping/view/MainViewModel.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.view
 
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.utils.MutableSingleLiveData
-import woowacourse.shopping.utils.SingleLiveData
+import woowacourse.shopping.utils.livedata.MutableSingleLiveData
+import woowacourse.shopping.utils.livedata.SingleLiveData
 
 class MainViewModel : ViewModel() {
     private val _updateProductEvent: MutableSingleLiveData<Map<Long, Int>> = MutableSingleLiveData()
